Create Snort Hands-On Lab
For this challenge, I needed to launch Snort (IDS/IPS) in sniffer mode to identify the attack source, service, and port involved in a brute-force attack. After analyzing the network traffic, I wrote an IPS rule to block the malicious activity, ensuring to first test it in "-A console" mode. Once verified, I ran Snort in IPS mode using "-A full" mode, blocking the malicious traffic for a minute. This process successfully stopped the attack, and as a result, a flag file appeared on the desktop, confirming the task was completed.